I designed a Space Shooter Game in 5 different Environment
The Space Shooter Game I designed and coded for has the following features.
1 - In the Console mode and MkeyKernel Mode section, there are multiple
enemies that come to my game from random parts of the screen. And the
spaceship has multiple fire features to hit these enemies. To give the
impression of space to my game, I visualized my game with a star effect.
2 - In the game I designed in BaseKernel, OpenGl and WINBGIM environments,
the multi-enemy and multiple bullet features are supported in the same way,
while on the other hand, I increased the movement speed of the enemies by 1
for every 5 targets hit.
3 - In all 5 environments, there is an End of Game screen and score information
written on this screen.